requirements.txt
boto3
requests
pytest
kafka-python==2.0.2
<<<<<<< HEAD
psycopg2-binary==2.9.9
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
sqlalchemy==2.0.23
requests==2.31.0
=======
faker==19.3.1
psycopg2-binary==2.9.7
pandas==2.0.3
plotly==5.16.1
streamlit==1.26.0
sqlalchemy==2.0.20
>>>>>>> 3d90db502757c7c66da83cc1a2e275b74dc59778
